move completed atom map builder
