+++ modules/sequence/src/main/scala/lucuma/odb/sequence/data/CompletedAtomMap.scala
@@ -3,9 +3,12 @@
 
 package lucuma.odb.sequence.data
 
+import cats.syntax.eq.*
 import cats.syntax.option.*
+import eu.timepit.refined.types.numeric.NonNegShort
 import eu.timepit.refined.types.numeric.PosInt
 import lucuma.core.enums.SequenceType
+import lucuma.core.model.sequence.Atom
 import lucuma.core.model.sequence.StepConfig
 
 /**
@@ -81,6 +84,53 @@ object CompletedAtomMap {
     ): (CompletedAtomMap[D], Boolean) = {
       val k = Key.fromProtoAtom(s, a)
       if (contains(k)) (decrement(k), true) else (m, false)
+    }
+
+  }
+
+  trait Builder[D] {
+    def reset: Builder[D]
+    def next(aid: Atom.Id, count: NonNegShort, sequenceType: SequenceType, step: StepMatch[D]): Builder[D]
+    def build: CompletedAtomMap[D]
+  }
+
+  object Builder {
+
+    def init[D]: Builder[D] = Reset(CompletedAtomMap.Empty)
+
+    private case class Reset[D](completed: CompletedAtomMap[D]) extends Builder[D] {
+      override def reset: Builder[D] = this
+
+      override def next(aid: Atom.Id, count: NonNegShort, sequenceType: SequenceType, step: StepMatch[D]): Builder[D] =
+        InProgress(aid, count, sequenceType, List(step), completed)
+
+      override def build: CompletedAtomMap[D] = completed
+    }
+
+    private case class InProgress[D](
+      inProgressAtomId:       Atom.Id,
+      inProgressCount:        NonNegShort,
+      inProgressSequenceType: SequenceType,
+      inProgressSteps:        AtomMatch[D],
+      completed:              CompletedAtomMap[D]
+    ) extends Builder[D] {
+
+      override def reset: Builder[D] = Reset(completed)
+
+      private def addStep(step: StepMatch[D]): InProgress[D] =
+        copy(inProgressSteps = step :: inProgressSteps)
+
+      override def next(aid: Atom.Id, count: NonNegShort, sequenceType: SequenceType, step: StepMatch[D]): Builder[D] =
+        if (aid === inProgressAtomId) addStep(step) // continue existing atom
+        else InProgress(aid, count, sequenceType, List(step), build) // start a new atom
+
+      private def inProgressKey: CompletedAtomMap.Key[D] =
+        CompletedAtomMap.Key(inProgressSequenceType, inProgressSteps.reverse)
+
+      override def build: CompletedAtomMap[D] =
+        if (inProgressSteps.sizeIs != inProgressCount.value) completed
+        else completed.increment(inProgressKey)
+
     }
 
   }
